sprint1/fast_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

psycopg2-binary==2.9.6
psycopg2-pool==1.1

pydantic==1.10.7

pyyaml==6.0

pytest==7.3.1
httpx==0.24.1
pytest-asyncio==0.21.0

gunicorn==20.1.0